+++ GiDapper/Program.cs
@@ -23,7 +23,7 @@ namespace GiDapper
             Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Revisiones());
+            Application.Run(new Clientes());
         }
     }
 }
